Scrambling Code is added
This code will be the workhorse of the application. I built it in
eclipse, independent of an Android design, so the code will have to be
refactored slightly in order to work in the app. Regardless, you should
be able to download it and use it to test it out. Remember, the SRC
contains the java code and the BIN contains the compiled java bytecode